Putting into Git
This is, as of now, proof of concept - a WPF translation of the Zone
Simulator that was made in GM8. This project simulates some of the floor
(around upto 30 floor numbers). The graphic locations are hard coded,
they must be in drive a:\ for this project to compile out of the box.